Move AlignAndFocusPowderSlim component headers to inc directory
