+++ Framework/DataHandling/inc/MantidDataHandling/AlignAndFocusPowderSlim/NexusLoader.h
